External Apps Need Refactor

+++ lib/subghz/protocols/legrand.h
@@ -82,7 +82,7 @@ void subghz_protocol_decoder_legrand_feed(void* context, bool level, uint32_t du
  * @param context Pointer to a SubGhzProtocolDecoderLegrand instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_legrand_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_legrand_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderLegrand.
